Done make all calculation average

+++ app/Http/Controllers/ManageCriteriaController.php
@@ -3,7 +3,11 @@
 namespace App\Http\Controllers;
 
 use App\Models\ManageCriteria;
+use App\Models\Secretary;
+use App\Models\SportsExco;
+use App\Models\StudentWarefare;
 use App\Models\VoterRate;
+use App\Models\WelfareExco;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
@@ -24,17 +28,93 @@ class ManageCriteriaController extends Controller
     {
         ManageCriteria::create($request->all());
         VoterRate::create($request->all());
+        StudentWarefare::create($request->all());
+        WelfareExco::create($request->all());
+        SportsExco::create($request->all());
+        Secretary::create($request->all());
         return redirect('/admin/viewCriteria');
     }
 
     // Update Criteria inside DB
     public function update(Request $request, $id)
     {
+        // Find and delete ManageCriteria
         $criteria = ManageCriteria::find($id);
-        $voter = VoterRate::find($id);
-        $criteria->update($request->all());
-        $voter->update($request->all());
-        return redirect('/admin/viewCriteria');
+
+        if ($criteria) {
+            $criteria->update($request->all());
+
+            // Find and delete associated VoterRate Database instances
+            $voter = VoterRate::where('criteria_id', 0);
+            $voterId = VoterRate::where('criteria_id', $id);
+
+            // Find and delete associated Student Warefare Database instances
+            $studentWarefare = StudentWarefare::where('criteria_id', 0);
+            $studentWarefareId = StudentWarefare::where('criteria_id', $id);
+
+            // Find and delete associated Welfare Exco Database instances
+            $welfareExco = WelfareExco::where('criteria_id', 0);
+            $welfareExcoId = WelfareExco::where('criteria_id', $id);
+
+            // Find and delete associated Sports Exco Database instances
+            $sportsExco = SportsExco::where('criteria_id', 0);
+            $sportsExcoId = SportsExco::where('criteria_id', $id);
+
+            // Find and delete associated Secretary Database instances
+            $secretary = Secretary::where('criteria_id', 0);
+            $secretaryId = Secretary::where('criteria_id', $id);
+
+            if ($voter && $studentWarefare && $welfareExco && $sportsExco && $secretary) {
+                $voter->update([
+                    'criteria_id' => $id,
+                    'name' => $request->input('name'),
+                ]);
+                $studentWarefare->update([
+                    'criteria_id' => $id,
+                    'name' => $request->input('name'),
+                ]);;
+                $welfareExco->update([
+                    'criteria_id' => $id,
+                    'name' => $request->input('name'),
+                ]);;
+                $sportsExco->update([
+                    'criteria_id' => $id,
+                    'name' => $request->input('name'),
+                ]);;
+                $secretary->update([
+                    'criteria_id' => $id,
+                    'name' => $request->input('name'),
+                ]);;
+            }
+
+            if ($voterId && $studentWarefareId && $welfareExcoId && $sportsExcoId && $secretaryId) {
+                $voterId->update([
+                    'criteria_id' => $id,
+                    'name' => $request->input('name'),
+                ]);;
+                $studentWarefareId->update([
+                    'criteria_id' => $id,
+                    'name' => $request->input('name'),
+                ]);;
+                $welfareExcoId->update([
+                    'criteria_id' => $id,
+                    'name' => $request->input('name'),
+                ]);;
+                $sportsExcoId->update([
+                    'criteria_id' => $id,
+                    'name' => $request->input('name'),
+                ]);;
+                $secretaryId->update([
+                    'criteria_id' => $id,
+                    'name' => $request->input('name'),
+                ]);
+            }
+
+            return redirect('/admin/viewCriteria');
+        } else {
+            // Handle case where ManageCriteria with $id is not found
+            return redirect('/admin/viewCriteria');
+        }
     }
 
     // Delete Criteria inside DB
@@ -46,16 +126,40 @@ class ManageCriteriaController extends Controller
         if ($criteria) {
             $criteria->delete();
 
-            // Find and delete associated VoterRate instances
+            // Find and delete associated VoterRate Database instances
             $voter = VoterRate::where('criteria_id', 0);
             $voterId = VoterRate::where('criteria_id', $id);
 
-            if ($voter) {
+            // Find and delete associated Student Warefare Database instances
+            $studentWarefare = StudentWarefare::where('criteria_id', 0);
+            $studentWarefareId = StudentWarefare::where('criteria_id', $id);
+
+            // Find and delete associated Welfare Exco Database instances
+            $welfareExco = WelfareExco::where('criteria_id', 0);
+            $welfareExcoId = WelfareExco::where('criteria_id', $id);
+
+            // Find and delete associated Sports Exco Database instances
+            $sportsExco = SportsExco::where('criteria_id', 0);
+            $sportsExcoId = SportsExco::where('criteria_id', $id);
+
+            // Find and delete associated Secretary Database instances
+            $secretary = Secretary::where('criteria_id', 0);
+            $secretaryId = Secretary::where('criteria_id', $id);
+
+            if ($voter && $studentWarefare && $welfareExco && $sportsExco && $secretary) {
                 $voter->delete();
+                $studentWarefare->delete();
+                $welfareExco->delete();
+                $sportsExco->delete();
+                $secretary->delete();
             }
 
-            if ($voterId) {
+            if ($voterId && $studentWarefareId && $welfareExcoId && $sportsExcoId && $secretaryId) {
                 $voterId->delete();
+                $studentWarefareId->delete();
+                $welfareExcoId->delete();
+                $sportsExcoId->delete();
+                $secretaryId->delete();
             }
 
             return redirect('/admin/viewCriteria')->with('delete', 'Your Card Success been Deleted!.');
